Avance semana X: CRUD funcional con conexión a base de datos

+++ delete.php
@@ -1,7 +1,14 @@
 <?php
 include('db.php');
-$id = $_GET['id'];
+
+if (!isset($_GET['id'])) {
+    header("Location: index.php");
+    exit();
+}
+
+$id = (int)$_GET['id'];
 mysqli_query($conn, "DELETE FROM tareas WHERE id=$id");
 header("Location: index.php");
 exit();
 ?>
+
